move like function, add heiht fit for tweet and comment
