+++ app/assets/stylesheets/comments.scss
@@ -47,6 +47,7 @@
 
         .twite {
           width: 80%;
+          height: fit-content;
           margin: 3px;
           background-color: beige;
           border-radius: 10px;
