+++ app/assets/stylesheets/profiles.scss
@@ -134,6 +134,7 @@
 
         .twite {
           width: auto;
+          height: fit-content;
           margin: 3px;
           background-color: beige;
           border-radius: 10px;
@@ -151,6 +152,7 @@
 
         .comment-text {
           width: auto;
+          height: fit-content;
           margin: 3px;
           background-color: thistle;
           border-radius: 10px;
