+++ app/assets/stylesheets/tweets.scss
@@ -44,6 +44,7 @@
 
         .twite {
           width: auto;
+          height: fit-content;
           margin: 3px;
           background-color: beige;
           border-radius: 10px;
@@ -216,6 +217,7 @@
           }
 
         .twite {
+          height: fit-content;
           margin: 3px;
           background-color: thistle;
           border-radius: 10px;
